Make admin.js a template
Make admin.js a template, so that we can interpolate the value of
the ADMIN_PUBLIC_URL parameter, to make the AJAX call for getting
the default email text to be sent to a user when their request is
to be hidden work correctly even when the admin site is served
through a secure proxy (as it is at mySociety).

Also remove some residual example code that presumably persists from
the demo theme this theme was derived from.